requirements.txt
streamlit>=1.31.0
pandas>=2.2.0
openpyxl>=3.1.2
streamlit-searchbox>=0.1.13
pymongo[srv]>=4.6.0
python-dotenv>=1.0.0
